Update wallet balance type to numeric with precision 2

+++ migrations/1710870278_init_schema.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "wallets" (
   "id" varchar PRIMARY KEY,
   "account_id" varchar NOT NULL,
-  "balance" bigint NOT NULL,
+  "balance" numeric(22, 2) NOT NULL,
   "type" varchar NOT NULL,
   "created_at" timestamptz NOT NULL DEFAULT 'now()',
   "updated_at" timestamptz NOT NULL DEFAULT 'now()'
